+++ src/test/java/com/withus/withmebe/comment/controller/CommentControllerTest.java
@@ -34,7 +34,7 @@ import org.springframework.http.MediaType;
 import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors;
 import org.springframework.test.web.servlet.MockMvc;
-import util.WithMockCustomUser;
+import util.security.WithMockCustomUser;
 
 
 @WebMvcTest(controllers = CommentController.class, excludeFilters = {
